requirements.txt
asgiref==3.7.1
backports.zoneinfo==0.2.1
Django==4.2.1
psycopg==3.1.9
psycopg2==2.9.6
sqlparse==0.4.4
typing_extensions==4.6.1
drf-yasg==1.21.5
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
django-cors-headers==4.0.0